docs: remove issues list and issue creation link
